Refactor payment and order item handling
Changed `amount` property in `PaymentRequest.cs` from `int` to `long` to support larger values. Updated `finalAmount` parsing in `ZarinpalFactory.cs` accordingly. Removed product category forms in `Create.cshtml` and `Edit.cshtml`. Enhanced `Index.cshtml` with namespace addition, label update, formatting changes, button text update, and URL parameter change. Added `OnGetItems` method in `Index.cshtml.cs` for partial view of order items. Implemented `GetItems` method in `IOrderApplication.cs`, `OrderApplication.cs`, `IOrderRepository.cs`, and `OrderRepository.cs` to fetch order items. Added `Items.cshtml` for displaying order items and `OrderItemViewModel.cs` for defining the view model.

+++ 0_Framework/Application/ZarinPal/ZarinpalFactory.cs
@@ -22,7 +22,7 @@ namespace _0_Framework.Application.ZarinPal
             long orderId)
         {
             amount = amount.Replace(",", "");
-            var finalAmount = int.Parse(amount);
+            var finalAmount = long.Parse(amount);
             var siteUrl = _configuration.GetSection("payment")["siteUrl"];
 
             var client = new RestClient($"https://{Prefix}.zarinpal.com/pg/v4/payment/request.json");
